Title: 앞의 두 수를 더한 수열 2, Time: 96ms, Memory: 8MB, Status: Wrong Answer - CodeTree

+++ 240311/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.java
@@ -8,7 +8,7 @@ public class Main {
 
         int n = Integer.parseInt(br.readLine());
 
-        int[] arr = new int[n + 3];
+        double[] arr = new double[n + 3];
 
         arr[1] = 1;
         arr[2] = 1;
